Move printer logic out of printer interface
Originally it was thought that most printers would execute a similar run
loop with reader and writer threads and possibly a print thread when
required. However it is possible in the future that printers will print
quite differently, particularly if they print from SD card.

As such move all of the runtime logic into the marlin printer for now.
Later if this is common to other serial printers it can be a separate
base class to be used in multiple inheritence.